// types.ts
export type UserRole = 'ADMIN' | 'OPERATOR' | 'DRIVER' | 'CUSTOMER';
export type UserStatus = 'PENDING_APPROVAL' | 'APPROVED' | 'SUSPENDED';
export type PaymentMethod = 'CASH' | 'WALLET';
export type VehicleType = 'TOKTOK' | 'MOTORCYCLE' | 'CAR';
// ...
export interface User {
  id: string;
  email: string;
  phone: string;
  name: string;
  role: UserRole;
  status: UserStatus;
  photoURL?: string;
  password?: string; // حقل كلمة السر (لأغراض الإدارة)
  vehicleType?: VehicleType;
  plateNumber?: string; // رقم لوحة المركبة
  operatorId?: string;
  zoneId?: string;
  wallet: {
    balance: number;
    totalEarnings: number;
    withdrawn: number;
  };
  location?: {
    lat: number;
    lng: number;
    updatedAt: number;
  };
}